+++ src/core_modules/capture-core/components/DataEntries/EnrollmentRegistrationEntry/hooks/useBuildEnrollmentPayload.js
@@ -10,7 +10,6 @@ import { convertClientToServer, convertFormToClient } from '../../../../converte
 import {
     convertDataEntryValuesToClientValues,
 } from '../../../DataEntry/common/convertDataEntryValuesToClientValues';
-import { capitalizeFirstLetter } from '../../../../../capture-core-utils/string';
 import { generateUID } from '../../../../utils/uid/generateUID';
 import {
     useBuildFirstStageRegistration,
@@ -25,8 +24,8 @@ import {
     deriveAutoGenerateEvents,
     deriveFirstStageDuringRegistrationEvent,
 } from '../../../Pages/New/RegistrationDataEntry/helpers';
-import { FEATURETYPE } from '../../../../constants';
 import type { EnrollmentPayload } from '../EnrollmentRegistrationEntry.types';
+import { geometryType, getPossibleTetFeatureTypeKey, buildGeometryProp } from '../../common/TEIAndEnrollment/geometry';
 
 type DataEntryReduxConverterProps = {
     programId: string;
@@ -66,25 +65,6 @@ function getServerValuesForMainValues(
     return serverValues;
 }
 
-function getPossibleTetFeatureTypeKey(serverValues: Object) {
-    return Object
-        .keys(serverValues)
-        .find(key => key.startsWith('FEATURETYPE_'));
-}
-
-function buildGeometryProp(key: string, serverValues: Object) {
-    if (!serverValues[key]) {
-        return undefined;
-    }
-    const type = capitalizeFirstLetter(key.replace('FEATURETYPE_', '').toLocaleLowerCase());
-    return {
-        type,
-        coordinates: serverValues[key],
-    };
-}
-
-const geometryType = formValuesKey => Object.values(FEATURETYPE).find(geometryKey => geometryKey === formValuesKey);
-
 const deriveAttributesFromFormValues = (formValues = {}) =>
     Object.keys(formValues)
         .filter(key => !geometryType(key))
@@ -116,7 +96,7 @@ export const useBuildEnrollmentPayload = ({
             dataEntryFieldsMeta,
             formFoundation,
         );
-        const { enrolledAt, occurredAt, assignee } = serverValuesForMainValues;
+        const { enrolledAt, occurredAt, assignee, geometry: enrollmentGeometry } = serverValuesForMainValues;
 
         const { stages } = getTrackerProgramThrowIfNotFound(programId);
 
@@ -167,21 +147,20 @@ export const useBuildEnrollmentPayload = ({
             enrolledAt,
             attributes,
             events: allEventsToBeCreated,
+            geometry: enrollmentGeometry,
         };
 
-        const tetFeatureTypeKey = getPossibleTetFeatureTypeKey(serverValuesForFormValues);
-        let geometry;
-        if (tetFeatureTypeKey) {
-            geometry = buildGeometryProp(tetFeatureTypeKey, serverValuesForFormValues);
-            delete serverValuesForFormValues[tetFeatureTypeKey];
-        }
+        const tetFeatureTypeKey = getPossibleTetFeatureTypeKey(formServerValues);
+        const tetGeometry = tetFeatureTypeKey ?
+            buildGeometryProp(tetFeatureTypeKey, formValues)
+            : undefined;
 
         return {
             trackedEntity: teiId || generateUID(),
             orgUnit: orgUnitId,
             trackedEntityType: trackedEntityTypeId,
             attributes,
-            geometry,
+            geometry: tetGeometry,
             enrollments: [enrollment],
         };
     };