+++ src/core_modules/capture-core/components/WidgetProfile/hooks/useTrackedEntityInstances.js
@@ -83,7 +83,7 @@ export const useTrackedEntityInstances = (
     }, [data?.trackedEntityInstance?.trackedEntityType, refetchTET]);
 
     useEffect(() => {
-        if (storedGeometry) {
+        if (storedGeometry !== undefined) {
             setGeometry(storedGeometry);
         }
     }, [storedGeometry]);