+++ src/core_modules/capture-core/components/WorkingLists/TeiWorkingLists/epics/teiViewEpics/helpers/getTeiListData/index.js
@@ -1,2 +1,2 @@
 // @flow
-export { getTeiListData } from './getTeiListData';
+export { getTeiListData, createApiQueryArgs as createApiTrackedEntitiesQueryArgs } from './getTeiListData';