+++ src/core_modules/capture-core/components/WorkingLists/TeiWorkingLists/epics/teiViewEpics/index.js
@@ -7,3 +7,4 @@ export {
     updateProgramStageTemplateEpic,
 } from './programStageTemplates.epics';
 export { retrieveAllTemplatesEpic, retrieveTEITemplatesEpic } from './retrieveTemplates.epics';
+export { createApiTrackedEntitiesQueryArgs } from './helpers';