+++ src/core_modules/capture-core/components/WorkingLists/WorkingListsBase/index.js
@@ -52,4 +52,5 @@ export type {
     UpdateTemplate,
     WorkingListTemplate,
     WorkingListTemplates,
+    WorkingListsOutputProps,
 } from './workingListsBase.types';