+++ src/core_modules/capture-core/components/WorkingLists/WorkingListsCommon/index.js
@@ -22,3 +22,4 @@ export type {
     Callbacks,
 } from './types';
 export { TEMPLATE_SHARING_TYPE } from './constants';
+export { DownloadDialog } from './DownloadDialog';