+++ CHANGELOG.md
@@ -1,3 +1,32 @@
+## [100.67.3](https://github.com/dhis2/capture-app/compare/v100.67.2...v100.67.3) (2024-03-20)
+
+
+### Bug Fixes
+
+* [DHIS2-17108][DHIS2-15681] enrollment and TET geometry are not stored and properly displayed ([#3582](https://github.com/dhis2/capture-app/issues/3582)) ([46b9546](https://github.com/dhis2/capture-app/commit/46b9546235c5fceafd8576af8c77687eeb61bd1a))
+
+## [100.67.2](https://github.com/dhis2/capture-app/compare/v100.67.1...v100.67.2) (2024-03-20)
+
+
+### Bug Fixes
+
+* [DHIS2-16767] Refactor widgets empty text to be dynamic ([#3574](https://github.com/dhis2/capture-app/issues/3574)) ([093cf9d](https://github.com/dhis2/capture-app/commit/093cf9d05e764eb4cccceb0e4a5b646308263b04))
+
+## [100.67.1](https://github.com/dhis2/capture-app/compare/v100.67.0...v100.67.1) (2024-03-19)
+
+
+### Bug Fixes
+
+* [DHIS2-16393] use option name in Stages&Events ([#3578](https://github.com/dhis2/capture-app/issues/3578)) ([b75398d](https://github.com/dhis2/capture-app/commit/b75398d605d5bee1c37e1a0c3f8718513a4aeb9c))
+* [DHIS2-16394] use option name in relationship table ([#3579](https://github.com/dhis2/capture-app/issues/3579)) ([6b98c37](https://github.com/dhis2/capture-app/commit/6b98c375bb782045ee204417e8169a264762191d))
+
+# [100.67.0](https://github.com/dhis2/capture-app/compare/v100.66.6...v100.67.0) (2024-03-19)
+
+
+### Features
+
+* [DHIS2-13233] download tracked entities working lists ([#3569](https://github.com/dhis2/capture-app/issues/3569)) ([cca2064](https://github.com/dhis2/capture-app/commit/cca2064f99c64c291da9a8e157948cc18cac7e43))
+
 ## [100.66.6](https://github.com/dhis2/capture-app/compare/v100.66.5...v100.66.6) (2024-03-18)